requirements.txt
streamlit
streamlit-option-menu
pymongo
hugchat
python-dotenv
transformers[torch]